Testing IMU Accelerometer and Gyroscope
Intended for use with Arduino Nano 33 BLE Sense, this code is a modification of the accelerometer testing file. This reads both the accelerometer and gyroscope data and prints it both to the serial monitor and to a predefined file in the attached SD card. If this is done successfully, the RGB LED shows green, otherwise, an error message is displayed on the serial monitor and the LED shows Red on the arduino